==> 2.3_Using_Software_to_Generate_Valid_Inferences.rmd <==
```{r, include=FALSE}
knitr::opts_chunk$set(warning = FALSE, 
                      message = FALSE,
                      cache = TRUE, 
                      cache.path = "cache/2.3/")
```



# Accounting for the Sampling Design {-}

The Ethiopia Socioeconomic Survey (ESS4) is part of the Living Standards Measurement Study (LSMS), an initiative by the World Bank in collaboration with the Central Statistical Agency of Ethiopia (CSA). Its objective is to provide high-quality data for the analysis of poverty, food security, agricultural development, and other key socioeconomic aspects.

The ESS4 is part of a series of longitudinal surveys initiated in 2011 under the name Ethiopia Rural Socioeconomic Survey (ERSS). In its first two rounds, the survey only covered rural areas and small towns, but starting from the second wave (ESS2, 2013/2014), the coverage was expanded to all urban areas, ensuring nationally representative estimates. The ESS1, ESS2, and ESS3 waves constitute Panel I, while ESS4 initiates a new panel, serving as the baseline for future surveys.

The survey provides information on household structure, employment, income, consumption, education, health, access to infrastructure, and characteristics of the agricultural sector. One of its key features is the combination of household survey data with administrative records and remote sensing technologies, such as satellite imagery, to improve the accuracy of economic and social indicator estimates.

## Sample design of the ESS4 2018/2019 {-}

The sampling design of ESS4 is based on a two-stage stratified probabilistic sampling, with different strategies for rural and urban areas. In the first stage, **enumeration areas (EA)** in rural zones were selected using **simple random sampling (SRS)** from a subsample of the 2018 Large-Scale Agricultural Survey (AgSS). In urban areas, EAs were directly selected from the urban frame using **systematic sampling with probability proportional to size (PPS)** within each region. This approach ensured a proportional allocation of the sample between urban and rural areas.  

In the second stage, households within each selected EA were chosen through **systematic random sampling**. In rural areas, **10 agricultural households** were selected from the AgSS sample, along with **2 non-agricultural households**, ensuring that the total number of agricultural households remained 10 when fewer than 2 non-agricultural households were available. In urban areas, **15 households per EA** were selected, regardless of economic activity. A total of **7,527 households across 565 EAs** were initially planned for the survey; however, due to security restrictions, the final number of completed surveys was **6,770 households across 535 EAs**. This design ensures representative coverage of socioeconomic conditions in both rural and urban areas.

## Data Source {-}  

The **ESS4** was implemented by the **CSA of Ethiopia**, in collaboration with the **World Bank** and other international organizations. Data collection was conducted through **direct interviews** with selected households, using **electronic devices** through the **Computer-Assisted Personal Interviewing (CAPI)** methodology.  

The database includes detailed information on various socioeconomic aspects, such as:  

- **Household structure and sociodemographic characteristics** (age, gender, education, health).  
- **Living conditions and economic well-being** (income, consumption, access to services).  
- **Housing characteristics** (construction materials, access to drinking water, sanitation, and electricity).  
- **Food security and agricultural production** (crops, livestock, access to markets).  
- **Access to financing and land tenure**.  
- **Mobility and migration** (internal displacement and job seeking).  


For more details on the methodology, sampling design, and data access, you can refer to the official source on the World Bank Microdata Library:  
[ESS4 Survey on Microdata Library - World Bank](https://microdata.worldbank.org/index.php/catalog/3823/study-description#study_desc1674579234511)

# 2.3 Using Software to Generate Valid Inferences {-}

This set of code blocks prepares the working environment, loads the main household dataset from ESS4, and generates summary tables with the number of households and enumeration areas by region and area type (urban/rural). It then defines the complex survey design, applying the appropriate weights and stratification. Lastly, it validates the structure of the design and presents a summary of its implementation, ensuring that the estimates are representative at both national and regional levels.



### Prepare the work environment {-}

This **R** code sets up a clean and optimized environment for data analysis before executing any processing. First, it removes all previously loaded objects from memory and performs garbage collection to free up resources, preventing potential conflicts with data from previous sessions. Then, it loads a set of essential libraries for data manipulation and analysis, including packages specialized in survey data, efficient handling of large datasets, and formatting of results. Additionally, it ensures that the `select` function from `dplyr` is explicitly used to avoid conflicts with other libraries.  

Finally, the code adjusts the available RAM memory limit, optimizing performance for large-scale data processing. This setup is particularly useful in survey analysis and statistical studies that involve working with extensive datasets and computationally intensive calculations.

```{r setup, include=TRUE, eval= TRUE, warning=FALSE, message=FALSE, error=FALSE}
#------------------------------------------------------------------------------#
#                           Cleaning R Environment                             #
#------------------------------------------------------------------------------#

rm(list = ls())
gc()

#------------------------------------------------------------------------------#
#                                Libraries                                     #
#------------------------------------------------------------------------------#

library(tidyverse)
library(survey)
library(srvyr)
library(data.table)
library(magrittr)
library(haven)
library(stringr)
library(tidyr)
library(knitr)
library(kableExtra)

# Ensure select function from dplyr is used explicitly
select <- dplyr::select

#------------------------------------------------------------------------------#
#                        Setting the RAM Memory Limit                          #
#------------------------------------------------------------------------------#

memory.limit(250000000)

```

### Loading Datasets {-}

This **R** code loads and prepares a dataset for the **ESS4** survey analysis. It uses the `read_sav()` function to import a **SPSS (.sav)** file, ensuring the correct reading of the database. Then, the `select()` function is applied to extract only the key variables, including unique household and enumeration area identifiers, geographic location information, and stratification into urban and rural areas. Additionally, the final adjusted sampling weight is included, which is essential for obtaining representative estimates.  


```{r load-household-data}
#------------------------------------------------------------------------------#
#                           Loading Datasets                                   #
#------------------------------------------------------------------------------#

# Household identification; location; household size and field staff identification 
HH_data <- read_sav("data/data_ESS4/sect_cover_hh_w4.sav") %>%
  select(household_id, # Unique Household Identifier
         ea_id,        # Unique Enumeration Area Identifier
         saq14,        # Area = Rural/Urban
         saq02,        # Zone Code
         saq01,        # Region Code
         pw_w4         # Final adjusted sampling weight
         )

```

This code generates two summary tables that count the **enumeration areas** and **households** by region and type of area (urban/rural).  

First, the **`ea_counts_by_region`** table calculates the number of **enumeration areas (EA)** in each region and zone. To achieve this, unique EAs are identified and grouped by region and area type (urban or rural). The structure is then reorganized using `pivot_wider()` to display the rural and urban categories as separate columns. Finally, the total number of EAs per region is calculated, and the columns are renamed for better clarity.  

On the other hand, the **`hh_counts_by_region`** table follows a similar process, but instead of counting EAs, it counts the number of **households** by region and area type. Unique households are extracted along with their respective region and zone, then grouped and counted by category. Subsequently, the data is reorganized using `pivot_wider()`, allowing urban and rural households to be distinguished in separate columns. Finally, the total number of households per region is calculated, and the columns are renamed to ensure a clear and consistent format.  


```{r count-eas}
ea_counts_by_region <- HH_data %>%
  distinct(ea_id, saq14, saq01) %>%
  mutate(saq14 = as_factor(saq14), saq01 = as_factor(saq01)) %>%
  group_by(saq01, saq14) %>%
  tally() %>%
  pivot_wider(
    names_from = saq14,
    values_from = n,
    values_fill = list(n = 0)
  ) %>%
  mutate(ea_counts_by_region = RURAL + URBAN) %>%
  rename_with( ~ paste0(., "_ea_id"), c("RURAL", "URBAN"))

# Count of Households (HH) by region and area (urban/rural) categories

hh_counts_by_region <- HH_data %>%
  distinct(household_id, saq01, saq14) %>%
  mutate(saq14 = as_factor(saq14), saq01 = as_factor(saq01)) %>%
  group_by(saq01, saq14) %>%
  tally() %>%
  pivot_wider(
    names_from = saq14,
    values_from = n,
    values_fill = list(n = 0)
  ) %>%
  mutate(TOTAL_HH = RURAL + URBAN) %>%
  rename_with( ~ paste0(., "_HH"), c("RURAL", "URBAN"))

# Merging both datasets on saq01
temp <- inner_join(hh_counts_by_region, ea_counts_by_region)

```

This **R** code generates a final summary of the survey results, including an additional row with the total consolidated for **"ETHIOPIA"**. To achieve this, it combines the regional data stored in `temp` with a new row representing the sum of all regions. This aggregation is performed using `summarise_if(is.numeric, sum)`, ensuring that all numeric variables are correctly summed, while the region column (`saq01`) is manually assigned as "ETHIOPIA".  

Next, the dataset is reorganized using `select()`, where key variables are renamed for better clarity. The resulting table includes the number of **rural and urban households**, the **number of enumeration areas** for each type of zone, and the **total counts** of households and enumeration areas. Finally, it is converted into a **data frame** to facilitate manipulation and visualization. This process provides a structured view of the distribution of the population and enumeration areas, allowing for both regional analysis and a consolidated national overview.

```{r count-households, eval=FALSE}
# Summarizing the results including an additional row for "ETHIOPIA"
survey_counts <- bind_rows(
  temp,
  temp %>% data.frame() %>% summarise_if(is.numeric, sum) %>%
    mutate(saq01 = "ETHIOPIA")
) %>%
  select(
    Region = saq01,
    RURAL_HH,
    RURAL_AE = RURAL_ea_id,
    URBAN_HH,
    URBAN_AE = URBAN_ea_id,
    TOTAL_HH,
    TOTAL_AE = ea_counts_by_region
  ) %>%
  data.frame()

# Display final result
data.frame(survey_counts)
```

The table summarizes the distribution of households and enumeration areas (EA) by region in Ethiopia, distinguishing between rural and urban areas. Nationally, the sample includes 6,770 households and 535 EAs, with a relatively balanced distribution between urban and rural areas.  

The regions with the highest number of rural households are Amhara (479), Oromia (453), and SNNP (422), while Addis Ababa is entirely urban (778 households and 52 EAs). In terms of EAs, Amhara and Oromia lead in rural areas, while Addis Ababa and Dire Dawa have higher urban concentrations.  


```{r count-households2, eval=TRUE, echo = FALSE}
# Summarizing the results including an additional row for "ETHIOPIA"
survey_counts <- bind_rows(
  temp,
  temp %>% data.frame() %>% summarise_if(is.numeric, sum) %>%
    mutate(saq01 = "ETHIOPIA")
) %>%
  select(
    Region = saq01,
    RURAL_HH,
    RURAL_AE = RURAL_ea_id,
    URBAN_HH,
    URBAN_AE = URBAN_ea_id,
    TOTAL_HH,
    TOTAL_AE = ea_counts_by_region
  ) %>%
  data.frame()

# Display final result
data.frame(survey_counts) %>% 
  kable( align = "c", format = "pipe")
```

### Defining the survey design {-}

This code defines the **survey design** for the **ESS4** survey using the `survey` package in **R**, ensuring that the analysis accurately reflects the sample structure.  

First, a new variable `strata` is created by combining the region (`saq01`) with the urban/rural classification (`saq14`), allowing for **proper stratification** of the design. Then, the dataset `HH_data` is converted into a survey design (`as_survey_design()`), specifying:  
- **`ids = ea_id`**: Identifier of the primary sampling unit (enumeration area, EA).  
- **`strata = strata`**: Stratification by region and urban/rural area.  
- **`weights = pw_w4`**: Use of the adjusted sampling weight to obtain representative estimates.  
- **`nest = TRUE`**: Indicates that sampling units are nested within strata.  


```{r}
#------------------------------------------------------------------------------#
# Defining the survey design
#------------------------------------------------------------------------------#

ESS4_design <- HH_data %>%
  mutate(strata = paste0(saq01, "_", saq14)) %>%
  as_survey_design(
    ids = ea_id,
    # Primary sampling unit identifier (EA)
    strata = strata,
    # Stratification by region (saq01) and urban/rural zone (saq14)
    weights = pw_w4,
    # Final adjusted weight
    nest = TRUE
  )
```

Additionally, the option `survey.lonely.psu = "fail"` is set, causing the code to produce an error if a stratum has only a single, non-certainty primary sampling unit (PSU), ensuring a correctly specified survey design. Finally, `summary(ESS4_design)` is used to display a summary of the design, allowing verification of its structure before analysis.

```{r, echo=TRUE,eval=TRUE}
# Makes it an error to have a stratum with a single, non-certainty PSU
options(survey.lonely.psu = "fail")
summary(ESS4_design)

```

The results confirm that the **ESS4** survey design was correctly implemented using a **stratified cluster sampling approach (PSU) with replacement**. The stratification combines the region and the urban/rural classification, ensuring representativeness at both levels.  

The sample includes **535 clusters (PSU)**, and the number of primary sampling units in each stratum matches the planned design, indicating a consistent execution. The selection probabilities vary by region and area type, reflecting a sampling design adjusted to the population’s characteristics.  

Overall, these results validate the sampling structure and ensure that the derived estimates are representative at both the national and regional levels, accounting for geographic distribution and urban-rural stratification.
